+++ test/test_docs.py
@@ -9,8 +9,7 @@ from pathlib import Path
 from unittest import TestCase
 
 from pydbml import PyDBML
-from pydbml.exceptions import ColumnNotFoundError
-from pydbml.exceptions import TableNotFoundError
+from pydbml.classes import Expression
 
 
 TEST_DOCS_PATH = Path(os.path.abspath(__file__)).parent / 'test_data/docs'
@@ -95,7 +94,8 @@ class TestDocs(TestCase):
         self.assertEqual([c.name for c in table.columns], ['id', 'username', 'full_name', 'gender', 'created_at', 'rating'])
         *_, gender, created_at, rating = table.columns
         self.assertEqual(gender.default, 'm')
-        self.assertEqual(created_at.default, '(now())')
+        self.assertIsInstance(created_at.default, Expression)
+        self.assertEqual(created_at.default.text, 'now()')
         self.assertEqual(rating.default, 10)
 
     def test_index_definition(self) -> None:
@@ -124,11 +124,17 @@ class TestDocs(TestCase):
         self.assertEqual(ix[4].subjects, [table['booking_date']])
         self.assertEqual(ix[4].type, 'hash')
 
-        self.assertEqual(ix[5].subjects, ['(id*2)'])
+        self.assertEqual(len(ix[5].subjects), 1)
+        self.assertIsInstance(ix[5].subjects[0], Expression)
+        self.assertEqual(ix[5].subjects[0].text, 'id*2')
 
-        self.assertEqual(ix[6].subjects, ['(id*3)', '(getdate())'])
+        self.assertEqual(len(ix[6].subjects), 2)
+        self.assertIsInstance(ix[6].subjects[0], Expression)
+        self.assertIsInstance(ix[6].subjects[1], Expression)
+        self.assertEqual(ix[6].subjects[0].text, 'id*3')
+        self.assertEqual(ix[6].subjects[1].text, 'getdate()')
 
-        self.assertEqual(ix[7].subjects, ['(id*3)', table['id']])
+        self.assertEqual(ix[7].subjects, [Expression('id*3'), table['id']])
 
     def test_relationships(self) -> None:
         results = PyDBML.parse_file(TEST_DOCS_PATH / 'relationships_1.dbml')